+++ data_aggregator/static/vue/home.js
@@ -7,8 +7,8 @@ import ActiveRangePicker from './components/home/active-range-picker.vue';
 import JobsTable from './components/home/jobs-table.vue';
 import JobsFilter from './components/home/jobs-filter.vue';
 import Chart from './components/home/chart.vue';
-import TimeDifferenceWidget from './components/home/time-difference-widget.vue';
-import TimeWidget from './components/home/time-widget.vue';
+import TimeDifferenceWidget from './components/home/widgets/time-difference-widget.vue';
+import TimeWidget from './components/home/widgets/time-widget.vue';
 
 Vue.component('jobs-table', JobsTable);
 Vue.component('jobs-filter', JobsFilter);
